new_flask/requirements.txt
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
keras==2.9.0
Keras-Preprocessing==1.1.2
Flask==2.1.2
Pillow==9.0.1
numpy==1.22.3
matplotlib==3.5.1
matplotlib-inline==0.1.2
sklearn==0.0
scikit-learn==1.0.2
pythonping==1.1.1